Set default return values

+++ common/src/main/java/us/myles/ViaVersion/protocols/protocol1_14to1_13_2/data/MappingData.java
@@ -47,7 +47,7 @@ public class MappingData {
             if (id == null) {
                 Via.getPlatform().getLogger().warning("Unknown blockstate " + key + " :(");
             } else {
-                MappingData.motionBlocking.add(id);
+                MappingData.motionBlocking.add(id.intValue());
             }
         }
